+++ packages/fast-tooling-react/src/form/form/form-item.select.tsx
@@ -24,10 +24,6 @@ class FormItemSelect extends FormItemBase<
      * Renders the component
      */
     public render(): JSX.Element {
-        if (this.props.options.length === 1 && this.props.required) {
-            return null;
-        }
-
         const value: any =
             typeof this.props.data !== "undefined"
                 ? this.props.data