+++ packages/fast-tooling-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-tooling-react@1.3.0...@microsoft/fast-tooling-react@1.3.1) (2019-04-26)
+
+**Note:** Version bump only for package @microsoft/fast-tooling-react
+
+
+
+
+
 # [1.3.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-tooling-react@1.2.5...@microsoft/fast-tooling-react@1.3.0) (2019-04-23)